split worker support windows.
